Changing fahr and celsius integer variables to floating point variables

+++ ch1/s1-2/temp_conv.c
@@ -4,7 +4,8 @@
 //
 int main(){
 
-	int fahr, celsius;
+//	int fahr, celsius;
+	float fahr, celsius;
 	int lower, upper, step;
 
 	step = 20; 	// Step size
@@ -14,9 +15,9 @@ int main(){
 	fahr = lower;
 	while(fahr <= upper){
 
-		celsius = 5 * (fahr - 32) / 9;
+		celsius = 5.0 * (fahr - 32.0) / 9.0;
 //		printf("%d\t%d\n", fahr, celsius);
-		printf("%3d %6d\n", fahr, celsius);
+		printf("%3f %15f\n", fahr, celsius);
 		fahr = fahr + step;
 	}
